refactor(backend): abstract away wca oauth
