+++ src/app/api/auth/wca/callback/route.ts
@@ -1,10 +1,11 @@
 import { cookies } from 'next/headers'
-import { wcaOauthClient } from '@/backend/auth/oauth'
-import { z } from 'zod'
+import {
+  getWcaClaims,
+  validateWcaAuthorizationCode,
+} from '@/backend/auth/oauth/wca'
 import { createUserAccount } from '@/backend/auth/user'
 import { auth } from '@/backend/auth'
 import { tryCatch } from '@/utils/try-catch'
-import { env } from '@/env'
 import { WCA_AUTH_SUCCESS_SEARCH_PARAM } from '../error-search-param'
 
 // TODO: conflict handling
@@ -27,10 +28,11 @@ export async function GET(request: Request): Promise<Response> {
     code === null ||
     state === null ||
     storedState === null ||
-    codeVerifier === null
+    codeVerifier === null ||
+    state !== storedState
   ) {
     console.error(
-      '[GOOGLE AUTH] no code/state/storedState/codeVerifier: ',
+      '[WCA] bad code/state/storedState/codeVerifier: ',
       code,
       state,
       storedState,
@@ -44,19 +46,12 @@ export async function GET(request: Request): Promise<Response> {
       },
     })
   }
-  if (state !== storedState) {
-    return new Response(null, {
-      status: 400,
-    })
-  }
 
-  const { data: tokenRaw, error: tokenError } = await tryCatch(
-    wcaOauthClient.validateAuthorizationCode(code, {
-      credentials: env.AUTH_WCA_CLIENT_SECRET,
-    }) as Promise<unknown>,
+  const { data: token, error: tokenError } = await tryCatch(
+    validateWcaAuthorizationCode(code),
   )
   if (tokenError) {
-    console.error('[WCA] invalid token schema: ', tokenError)
+    console.error('[WCA] token validation error', tokenError)
     redirectTo.searchParams.append(WCA_AUTH_SUCCESS_SEARCH_PARAM, 'false')
     return new Response(null, {
       status: 302,
@@ -66,17 +61,7 @@ export async function GET(request: Request): Promise<Response> {
     })
   }
 
-  const token = tokenSchema.parse(tokenRaw)
-
-  const { me: claims } = await fetch(
-    'https://www.worldcubeassociation.org/api/v0/me',
-    {
-      method: 'GET',
-      headers: { Authorization: `Bearer ${token.access_token}` },
-    },
-  )
-    .then((res) => res.json())
-    .then((json) => meSchema.parse(json))
+  const claims = await getWcaClaims({ access_token: token.access_token })
 
   await createUserAccount({
     userId: session.user.id,
@@ -96,46 +81,3 @@ export async function GET(request: Request): Promise<Response> {
     },
   })
 }
-
-const countrySchema = z.object({
-  id: z.string(),
-  name: z.string(),
-  // continentId: z.string(),
-  iso2: z.string(),
-})
-
-export const avatarSchema = z.object({
-  id: z.number(),
-  status: z.string(),
-  url: z.string().url(),
-  thumb_url: z.string().url(),
-  is_default: z.boolean(),
-  can_edit_thumbnail: z.boolean(),
-})
-
-export const meSchema = z.object({
-  me: z.object({
-    // id: z.number(),
-    // created_at: z.string().datetime(),
-    // updated_at: z.string().datetime(),
-    name: z.string(),
-    wca_id: z.string(),
-    gender: z.string(),
-    // url: z.string().url(),
-    country: countrySchema,
-    delegate_status: z.string().nullable(),
-    // class: z.string(),
-    // teams: z.array(z.any()), // Can be replaced with a stricter type if teams structure is known
-    avatar: avatarSchema,
-    email: z.string().email(),
-  }),
-})
-
-export const tokenSchema = z.object({
-  access_token: z.string(),
-  token_type: z.literal('Bearer'), // since it's always "Bearer"
-  expires_in: z.number(),
-  refresh_token: z.string(),
-  scope: z.string(),
-  created_at: z.number(), // Unix timestamp
-})
